requirements.txt
python-dotenv
aiogram==2.5
google-generativeai


